remove sessions2023 from dsnj export
