+++ admin/src/scenes/dsnj-export/index.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
 import { toastr } from "react-redux-toastr";
 import * as FileSaver from "file-saver";
-import { ROLES, sessions2023, translate } from "snu-lib";
+import { ROLES, translate, getCohortPeriod } from "snu-lib";
 import dayjs from "@/utils/dayjs.utils";
 
 import { Title } from "../plan-transport/components/commons";
@@ -14,23 +14,10 @@ import { useSelector } from "react-redux";
 
 const exportDateKeys = ["cohesionCenters", "youngsBeforeSession", "youngsAfterSession"];
 
-const cohortOptions = sessions2023.map(({ id, dateStart, dateEnd }) => {
-  const options = { year: "numeric", month: "long", day: "numeric" };
-  const formatter = new Intl.DateTimeFormat("fr-FR", options);
-  const formattedStart = formatter.format(dateStart);
-  const formattedEnd = formatter.format(dateEnd);
-
-  const formattedRange = `${formattedStart} - ${formattedEnd}`;
-
-  return {
-    label: `Séjour <b>${formattedRange}</b>`,
-    value: id,
-  };
-});
-
 const DSNJExport = () => {
-  const [cohortId, setCohortId] = useState(cohortOptions[0].value);
+  const [cohortId, setCohortId] = useState(null);
   const [cohorts, setCohorts] = useState([]);
+  const [cohortOptions, setCohortOptions] = useState([]);
   const [isLDownloadingByKey, setDownloadingByKey] = useState({});
   const user = useSelector((state) => state.Auth.user);
   useDocumentTitle("Export DSNJ");
@@ -42,24 +29,24 @@ const DSNJExport = () => {
   const getCohorts = async () => {
     const { data: cohortsWithExportDates, ok } = await api.get("/cohort");
     if (!ok) return toastr.error("Impossible de récupérer la date des exports", "");
-    const cohortList = sessions2023.map(({ id, dateEnd }) => {
-      let dsnjExportDates = {};
-      const sessionWithExportDates = cohortsWithExportDates.find((current) => id === current.snuId);
-      if (sessionWithExportDates) {
-        dsnjExportDates = sessionWithExportDates.dsnjExportDates || {};
-      }
-
+    const cohortList = cohortsWithExportDates.map(({ snuId, dateEnd, dsnjExportDates }) => {
       // export available until 1 month after the cohort
-      const exportsAvailableUntil = new Date(dateEnd.getTime());
+      const exportsAvailableUntil = new Date(dateEnd);
       exportsAvailableUntil.setMonth(exportsAvailableUntil.getMonth() + 1);
 
       return {
-        id,
-        dsnjExportDates,
+        id: snuId,
+        dsnjExportDates: dsnjExportDates || {},
         exportsAvailableUntil,
       };
     });
     setCohorts(cohortList);
+    const formattedCohortOptions = cohortsWithExportDates.map((cohort) => ({
+      label: `Séjour <b>${getCohortPeriod(cohort)}</b>`,
+      value: cohort.snuId,
+    }));
+    setCohortOptions(formattedCohortOptions);
+    setCohortId(formattedCohortOptions[0].value);
   };
 
   const updateExportDate = (key) => async (date) => {
