+++ packages/lib/sessions.js
@@ -99,4 +99,21 @@ const sessions2023 = [
   },
 ];
 
-export { sessions2023, oldSessions };
+const getCohortPeriod = (cohort) => {
+  const startDate = new Date(cohort.dateStart);
+  const endDate = new Date(cohort.dateEnd);
+  const endDateformatOptions = { year: "numeric", month: "long", day: "numeric" };
+  const startDateformatOptions = { day: "numeric" };
+  if (startDate.getMonth() !== endDate.getMonth()) {
+    startDateformatOptions.month = "long";
+  }
+  if (startDate.getFullYear() !== endDate.getFullYear()) {
+    startDateformatOptions.year = "numeric";
+  }
+  const formattedStart = new Intl.DateTimeFormat("fr-FR", startDateformatOptions).format(startDate);
+  const formattedEnd = new Intl.DateTimeFormat("fr-FR", endDateformatOptions).format(endDate);
+
+  return `du ${formattedStart} au ${formattedEnd}`;
+};
+
+export { sessions2023, oldSessions, getCohortPeriod };
